21 files changed, 290 insertions(+), 296 deletions(-)

+++ piscines/php/rle-ru/rush00/install.php
@@ -1,5 +1,11 @@
 <?php
-mkdir("../private");
-file_put_contents("../private/accounts", NULL);
-file_put_contents("../private/products", NULL);
+function install()
+{
+if (!file_exists("../private"))
+	mkdir("../private");
+if (!file_exists("../private/accounts"))
+	file_put_contents("../private/accounts", NULL);
+if (!file_exists("../private/products"))
+	file_put_contents("../private/products", NULL);
+}
 ?>
